+++ book_orm/src/main/java/ru/chaban/ex05/domain/Author.java
@@ -10,7 +10,6 @@ import java.util.List;
  */
 @Entity
 @Table(name = "authors")
-@Transactional
 public class Author {
 
     @Column(name = "name")
