+++ book_orm/src/main/java/ru/chaban/ex05/domain/Book.java
@@ -10,7 +10,6 @@ import java.util.List;
  */
 @Entity
 @Table(name = "books")
-@Transactional
 public class Book {
 
     @Column(name = "name")
