+++ book_orm/src/main/java/ru/chaban/ex05/domain/Genre.java
@@ -10,7 +10,6 @@ import java.util.List;
  */
 @Entity
 @Table(name = "genres")
-@Transactional
 public class Genre {
 
     @Column(name = "name")
