+++ book_orm/src/main/java/ru/chaban/ex05/repository/AuthorRepositoryJpa.java
@@ -10,11 +10,5 @@ public interface AuthorRepositoryJpa {
 
     Optional<Author> findById(long id);
 
-    List<Author> findAll();
-
-    List<Author> findByName(String name);
-
-    void updateNameById(long id, String name);
-
     void deleteById(long id);
 }
