+++ book_orm/src/main/java/ru/chaban/ex05/repository/BookRepositoryJpa.java
@@ -10,11 +10,5 @@ public interface BookRepositoryJpa {
 
     Optional<Book> findById(long id);
 
-    List<Book> findAll();
-
-    List<Book> findByName(String name);
-
-    void updateNameById(long id, String name);
-
     void deleteById(long id);
 }
