+++ book_orm/src/main/java/ru/chaban/ex05/repository/GenreRepositoryJpa.java
@@ -10,11 +10,5 @@ public interface GenreRepositoryJpa {
 
     Optional<Genre> findById(long id);
 
-    List<Genre> findAll();
-
-    List<Genre> findByName(String name);
-
-    void updateNameById(long id, String name);
-
     void deleteById(long id);
 }
