+++ book_orm/src/main/java/ru/chaban/ex05/repository/GenreRepositoryJpaImpl.java
@@ -11,7 +11,6 @@ import javax.persistence.TypedQuery;
 import java.util.List;
 import java.util.Optional;
 
-@Transactional
 @Repository
 public class GenreRepositoryJpaImpl implements GenreRepositoryJpa {
 
@@ -31,32 +30,6 @@ public class GenreRepositoryJpaImpl implements GenreRepositoryJpa {
     @Override
     public Optional<Genre> findById(long id) {
         return Optional.ofNullable(em.find(Genre.class, id));
-    }
-
-    @Override
-    public List<Genre> findAll() {
-        TypedQuery<Genre> query = em.createQuery("select s from Genre s ", Genre.class);
-        return query.getResultList();
-    }
-
-    @Override
-    public List<Genre> findByName(String name) {
-        TypedQuery<Genre> query = em.createQuery("select s " +
-                        "from Genre s " +
-                        "where s.name = :name",
-                Genre.class);
-        query.setParameter("name", name);
-        return query.getResultList();
-    }
-
-    @Override
-    public void updateNameById(long id, String name) {
-        Query query = em.createQuery("update Genre s " +
-                "set s.name = :name " +
-                "where s.id = :id");
-        query.setParameter("name", name);
-        query.setParameter("id", id);
-        query.executeUpdate();
     }
 
     @Override
