+++ book_orm/src/main/java/ru/chaban/ex05/repository/MyBooksRepositoryJpaImpl.java
@@ -12,7 +12,6 @@ import javax.persistence.TypedQuery;
 import java.util.List;
 import java.util.Optional;
 
-@Transactional
 @Repository
 public class MyBooksRepositoryJpaImpl implements MyBooksRepositoryJpa {
 
@@ -32,33 +31,6 @@ public class MyBooksRepositoryJpaImpl implements MyBooksRepositoryJpa {
     @Override
     public Optional<MyBooks> findById(long id) {
         return Optional.ofNullable(em.find(MyBooks.class, id));
-    }
-
-    @Override
-    public List<MyBooks> findAll() {
-        TypedQuery<MyBooks> query = em.createQuery("select s from MyBooks s", MyBooks.class);
-        return query.getResultList();
-    }
-
-    @Override
-    public List<MyBooks> findByName(String name) {
-        TypedQuery<MyBooks> query = em.createQuery("select s " +
-                        "from MyBooks s, Book b " +
-                        "where s.book = b.id and b.name = :name",
-                MyBooks.class);
-        query.setParameter("name", name);
-        return query.getResultList();
-    }
-
-    @Override
-    @Transactional
-    public void updateNameById(long id, Book book) {
-        Query query = em.createQuery("update MyBooks s " +
-                "set s.book = :book " +
-                "where s.id = :id");
-        query.setParameter("book", book);
-        query.setParameter("id", id);
-        query.executeUpdate();
     }
 
     @Override
