+++ book_orm/src/main/java/ru/chaban/ex05/repository/OpinionRepositoryJpaImpl.java
@@ -11,7 +11,6 @@ import javax.persistence.TypedQuery;
 import java.util.List;
 import java.util.Optional;
 
-@Transactional
 @Repository
 public class OpinionRepositoryJpaImpl implements OpinionRepositoryJpa {
 
@@ -31,32 +30,6 @@ public class OpinionRepositoryJpaImpl implements OpinionRepositoryJpa {
     @Override
     public Optional<Opinion> findById(long id) {
         return Optional.ofNullable(em.find(Opinion.class, id));
-    }
-
-    @Override
-    public List<Opinion> findAll() {
-        TypedQuery<Opinion> query = em.createQuery("select s from Opinion s", Opinion.class);
-        return query.getResultList();
-    }
-
-    @Override
-    public List<Opinion> findByName(String name) {
-        TypedQuery<Opinion> query = em.createQuery("select s " +
-                        "from Opinion s " +
-                        "where s.comment = :name",
-                Opinion.class);
-        query.setParameter("name", name);
-        return query.getResultList();
-    }
-
-    @Override
-    public void updateNameById(long id, String name) {
-        Query query = em.createQuery("update Opinion s " +
-                "set s.comment = :name " +
-                "where s.id = :id");
-        query.setParameter("name", name);
-        query.setParameter("id", id);
-        query.executeUpdate();
     }
 
     @Override
