+++ book_orm/src/main/java/ru/chaban/ex05/service/AuthorServiceImpl.java
@@ -1,0 +1,45 @@
+package ru.chaban.ex05.service;
+
+import org.springframework.stereotype.Repository;
+import ru.chaban.ex05.domain.Author;
+
+import javax.persistence.EntityManager;
+import javax.persistence.PersistenceContext;
+import javax.persistence.Query;
+import javax.persistence.TypedQuery;
+import java.util.List;
+import java.util.Optional;
+
+@Repository
+public class AuthorServiceImpl implements AuthorService {
+
+    @PersistenceContext
+    private EntityManager em;
+
+    @Override
+    public List<Author> findAll() {
+        TypedQuery<Author> query = em.createQuery("select s from Author s", Author.class);
+        return query.getResultList();
+    }
+
+    @Override
+    public List<Author> findByName(String name) {
+        TypedQuery<Author> query = em.createQuery("select s " +
+                        "from Author s " +
+                        "where s.name = :name",
+                Author.class);
+        query.setParameter("name", name);
+        return query.getResultList();
+    }
+
+    @Override
+    public void updateNameById(long id, String name) {
+        Query query = em.createQuery("update Author s " +
+                "set s.name = :name " +
+                "where s.id = :id");
+        query.setParameter("name", name);
+        query.setParameter("id", id);
+        query.executeUpdate();
+    }
+
+}
