+++ book_orm/src/main/java/ru/chaban/ex05/service/GenreService.java
@@ -1,0 +1,17 @@
+package ru.chaban.ex05.service;
+
+import ru.chaban.ex05.domain.Genre;
+
+import java.util.List;
+import java.util.Optional;
+
+public interface GenreService {
+
+
+    List<Genre> findAll();
+
+    List<Genre> findByName(String name);
+
+    void updateNameById(long id, String name);
+
+}
