+++ book_orm/src/main/java/ru/chaban/ex05/service/MyBooksService.java
@@ -1,0 +1,18 @@
+package ru.chaban.ex05.service;
+
+import ru.chaban.ex05.domain.Book;
+import ru.chaban.ex05.domain.MyBooks;
+
+import java.util.List;
+import java.util.Optional;
+
+public interface MyBooksService {
+
+
+    List<MyBooks> findAll();
+
+    List<MyBooks> findByName(String name);
+
+    void updateNameById(long id, Book book);
+
+}
